Refactor test helpers to separate package
Ensures that packages that are used only for tests are not
part of the module import graph.
